Add bindings to hb_ft_font_create_referenced.
This allows creating a harfbuzz font directly from a FreeType
face. Internally it increments the ref count of the FT face
so that there are no lifetime issues.

This allows much simpler use of the harfbuzz API rather than
having to implement all the custom callback types, when
using FT.

It's exposed on all non-mac UNIX platforms (Linux, BSD etc)
that are not Mac, and also on Android.