Fix pulldown / markdown warning with current rustc.
Due to the `#foo` syntax in the doc comments, if the `#` was at the
start of the line, the 2 different markdown implementations were
interpreting that differently, resulting in a new warning when
building the docs.

This changes all occurrences of this syntax, but doesn't address
broader issues of the markdown formatting.