Don't print cargo:include= with pkg-config
I introduced this in 692048e56a66afc36d4a6ef261ed2965bdcdf30a, but it's
the wrong thing to do. When a dependent crate gets the pkg-config
include paths in DEP_HARFBUZZ_INCLUDE, it may also need other aspects
such as the link paths, but it doesn't know that information is
available from pkg-config. So, the dependent crate must assume that, if
DEP_HARFBUZZ_INCLUDE is not available, it should use pkg-config.