ci: Rename Linux jobs correctly.
The shared job should not disable pkg-config usage and should
be the one that does a Harfbuzz build.

The static job should disable pkg-config and should not do
its own Harfbuzz build outside of the `harfbuzz-sys` build
scripts.

This is a precursor to making sure things are actually set up
correctly and clearly doing the appropriate thing.