Correct the name of the GitHub Actions jobs and make sure to test all static/shared combinations (#216)
* ci: Rename Linux jobs correctly.

The shared job should not disable pkg-config usage and should
be the one that does a Harfbuzz build.

The static job should disable pkg-config and should not do
its own Harfbuzz build outside of the `harfbuzz-sys` build
scripts.

This is a precursor to making sure things are actually set up
correctly and clearly doing the appropriate thing.

* ci: macOS-shared: Use harfbuzz from homebrew.

This is an experiment to see if we're able to pick it up from
pkg-config.