Do better with platform (font engine) features. (#217)
* freetype: Make optional, rename feature.

Now, the feature for this is called `freetype` rather than
`build-native-freetype`. This was never building FreeType
but just saying to build with FreeType and to use `freetype-sys`.

The default features were disabled for `freetype` and then enabling
the `freetype-sys` feature, but that's exactly what the `freetype`
crate does by default, so simplify by not trying to do it again.

Partial fix for #184 (this doesn't handle the `coretext` side).

* Add `directwrite` feature.

This controls whether or not `winapi` is used to support the
DirectWrite code on Windows.

* Add `coretext` feature.

This controls whether or not the CoreText APIs are supported
on Apple platforms.

Fixes #184.

* Add features list to crate docs and READMEs.

* Add notes to functions about feature required.