Use `bundled` feature to use our own copy of harfbuzz.
This looks for one installed on the system by default. It can be
forced to look on the system even if `bundled` is set by defining
`HARFBUZZ_SYS_USE_PKG_CONFIG` in the environment to `"1"`.

If `bundled` is set, then it will use the sources that we vendor.

In general, `bundled` should be set as needed by applications rather
than by libraries using `harfbuzz` or `harfbuzz-sys`. This allows
for control by the application's build since once the feature is
enabled by a crate, it can't readily be disabled.

The name for the `bundled` feature comes from the rusqlite bindings
for `libsqlite-sys`.