Fix `harfbuzz-sys` licensing.
While the top level files indicated that everything here was
dual licensed with `MIT OR Apache-2.0`, the actual `harfbuzz-sys`
crate's `Cargo.toml` claimed to be `MIT`. The source files in
the crate didn't indicate any licensing info.

This was an error dating back to
5c9adb40fa8bed0470b2c9951a1c65d588f9d8eb from 2015.

Prior to that commit, the bindings were clearly marked as being
`MIT OR Apache-2.0` and only the `Cargo.toml` licensing metadata
was incorrect after that.

In any event, the code here is almost entirely auto-generated by
`bindgen` and not something interesting.