ci: Build shared linux against system lib. (#228)
This stops building harfbuzz separately and tries just using
the system install to make sure that works.

This means that we build against an older version depending on
what's available on Ubuntu (since our macOS build uses homebrew
which tends to be fairly current).

Currently, that's version 2.7.4, the same as we're currently
building and packaging. Going forward, as we update what's in
use here, that should help us make sure that we're working with
older versions that can be found in the field, but we'll have to
take care to feature-gate newer APIs.